requirements.txt
python-docx>=1.1.0
lxml>=4.9.0
openai>=1.40.0
python-dotenv>=1.0.1
langgraph>=0.2.35
langchain>=0.3.8
langchain-openai>=0.2.9

